requirements.txt
pandas
flask
faiss-cpu
sentence-transformers
gunicorn